Moved TypeTagParser to its own file, added documentation for isStringTypeTag(), and replaced redundant Token types with documentation
